Changes and simplifications for creating the framework for an appropriate HNM time line format...
git-svn-id: https://mary.opendfki.de/repos/trunk@1935 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/marytts/signalproc/sinusoidal/hntm/analysis/FrameHarmonicPart.java
@@ -50,9 +50,9 @@ public class FrameHarmonicPart
         }
     }
     
-    public double[] getCeps(double f0InHz, int samplingRateInHz, HntmAnalyzerParams params)
+    public float[] getCeps(double f0InHz, int samplingRateInHz, HntmAnalyzerParams params)
     {
-        double[] ceps = null;
+        float[] ceps = null;
 
         if (complexAmps!=null)
         {
